Adding a hashmap to store the instances by the displayed text, to remove linear search over the whole list from helm-q-get-instance-by-display and thus a nested loop from helm-q-source-match-function
This fixes quadratic (approximately) execution time explosion with the number of instances in the list.

Right now the hashmap is used in addition to the `candidates' list: if the list is removed and a function is used to compute it from the hashmap, Helm calls the function on each character typed in the completion.